% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriDev.R
\name{as_triangle}
\alias{as_triangle}
\title{Convert a \code{data.table} object to \code{triangle}}
\usage{
as_triangle(DT, origin = 1L)
}
\description{
Convert a \code{data.table} object to \code{triangle}
}
\examples{
prem <- as_triangle(TRI$Premium)
loss <- as_triangle(TRI$Incurred)
loss / prem
}
