% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YearOccTable.R
\name{create_YOT}
\alias{create_YOT}
\title{Year Occurrence Table (YOT)}
\usage{
create_YOT(
  DT,
  year_id = "Year",
  occ_id = NULL,
  id_vars = NULL,
  value_vars = NULL
)
}
\description{
Year Occurrence Table (YOT)
}
