% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriDev.R
\name{dev_ChainLadder}
\alias{dev_ChainLadder}
\title{Complete the input triangle using ChainLadder method}
\usage{
dev_ChainLadder(tri, link_ratio = "weighted", floor = 1, tail_factor = 1)
}
\description{
Complete the input triangle using ChainLadder method
}
\examples{
loss <- as_triangle(TRI$Incurred)
dev_ChainLadder(loss)
}
