% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriDev.R
\name{dev_link_ratio}
\alias{dev_link_ratio}
\title{Calculate link ratio (i.e. age-to-age increment development factor)}
\usage{
dev_link_ratio(tri, style = c("weighted", "simple"))
}
\description{
Calculate link ratio (i.e. age-to-age increment development factor)
}
