% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{gMean}
\alias{gMean}
\alias{gSD}
\alias{gVaR}
\alias{gTVaR}
\title{Common summary statistics in reinsurance practice}
\usage{
gMean(x, ...)

gSD(x, ...)

gVaR(x, ...)

gTVaR(x, ...)
}
\description{
\code{gMean} mean
\code{gSD} standard deviation
\code{gVaR} Value at Risk
\code{gTVaR} Tail Value at Risk (a.k.a Conditional Tail Expectation, or CTE)
}
