% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_extra.R
\name{op_palette}
\alias{op_palette}
\alias{op_palette_plot}
\title{A selection of color palettes}
\usage{
op_palette(x = NULL, name = c("gc", "economist"))

op_palette_plot(p)
}
\description{
A selection of color palettes
}
\examples{
op_palette() |> op_palette_plot()
op_palette(c(1, 9)) |> op_palette_plot()
}
