% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opre-package.R
\docType{package}
\name{opre-package}
\alias{opre}
\alias{opre-package}
\title{opre: A package for reinsurance-related works}
\description{
A collection of functions and classes (mainly R6) for reinsurance analytical work (e.g. actuarial, modeling, data cleaning and etc.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/addnox/opre}
  \item Report bugs at \url{https://github.com/addnox/opre/issues}
}

}
\author{
\strong{Maintainer}: Charles Pan \email{addnox@gmail.com}

}
\keyword{internal}
