% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-ContractTerm.R
\name{parse_reinstatement}
\alias{parse_reinstatement}
\title{Parse reinstatement terms into data.frame}
\usage{
parse_reinstatement(reinstat)
}
\arguments{
\item{reinstat}{a character atomic variable, e.g. "2@100\\%"}
}
\description{
Parse reinstatement terms into data.frame
}
\examples{
parse_reinstatement("1@120\%, 2@100\%")
parse_reinstatement("1@120\%;2@100\%")
parse_reinstatement(c(1.2, 1))
parse_reinstatement(0) ## note this mean 1@0\%, rather than Nil reinstatement
parse_reinstatement("2@120\%")
parse_reinstatement("2@free")
parse_reinstatement("uf")
parse_reinstatement("nil")
parse_reinstatement("0@100\%")
}
