% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RePrice.R
\name{price_ImpliedLoading}
\alias{price_ImpliedLoading}
\title{Calculate loading factor}
\usage{
price_ImpliedLoading(
  ContractLoss,
  Premium,
  OtherDeduction = 0,
  prob = c(0.99, 0.995, 0.996)
)
}
\description{
Calculate loading factor
}
