% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractTerm.R
\name{term_Excess}
\alias{term_Excess}
\alias{term_Commission}
\alias{term_ProfitCommission}
\alias{term_LossParticipation}
\title{Common contract terms}
\usage{
term_Excess(
  Loss,
  Premium = NULL,
  Limit,
  Deductible,
  Participation = 1,
  is.FranchiseLimit = FALSE,
  is.FranchiseDed = FALSE,
  is.StopLoss = FALSE
)

term_Commission(Loss, Premium, Commission, right.closed = TRUE)

term_ProfitCommission(Loss, Premium, ProfitShare, ExpenseAllowance, digits = 2)

term_LossParticipation(
  Loss,
  Premium,
  Scheme,
  is.CedentPerspective = TRUE,
  digits = 2
)
}
\arguments{
\item{Loss, Premium}{Numeric vector for loss and premium}

\item{Limit, Deductible}{Limit and Deductible terms}

\item{Participation}{Used to multiply with final loss to layer}

\item{is.FranchiseLimit}{Logical, franchise limit means when LossAboveDeducible is greater than Limit, then no reimbursement}

\item{is.FranchiseDed}{Logical, franchise deductible means when Loss is greater than Deductible, full f.g.u. loss is applicable for reimbursement}

\item{is.StopLoss}{Logical, when TRUE, \code{Limit} and \code{Deductible} are deemed as in loss ratio form.  But \code{Loss} and final output are still be amount form}

\item{Commission}{A 3-column \code{data.frame}, with the sequence of \code{LowerLR}, \code{UpperLR} and \code{Commission}}

\item{right.closed}{Logical, default to be TRUE.  RightBound inclusive}

\item{ProfitShare}{The figure of \code{ProfitShare}.  Or a 3-column \code{data.frame}, with the sequence of \code{LowerProfit}, \code{UppoerProfit} and \code{ProfitShare}}

\item{ExpenseAllowance}{i.e. Management Expense}

\item{digits}{To round loss ratio percentage.  e.g. 54.1234\% will be rounded up to 54.12\% if \code{digits = 2}}

\item{Scheme}{A 3-column \code{data.frame}, with the sequence of \code{LowerLR}, \code{UpperLR} and \code{Participation}}

\item{is.CedentPerspective}{Logical.  If \code{TRUE}, the output is negative representing the loss amount that the cedent will bear}
}
\description{
\code{term_Excess} calculates LossInLayer based on specified Limit and Deductible structure

\code{term_Commission} calculates commssion figures based on given sliding scale table

\code{term_Reinstatement} returns a list of two vectors, one being reinstated loss and the other is reinstated premium

\code{term_ProfitCommission} calculates profit commission

\code{term_LossParcipation} calculates loss participation (i.e. loss corridor) from cedent's perspective
}
\details{
Many terms are plain-vanilla and are recommended to use base R calculations.  Only complicated terms are implemented.
}
