% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_extra.R
\name{theme_op}
\alias{theme_op}
\title{Basic ggplot2 theme}
\usage{
theme_op(
  base_size = 12,
  base_family = "",
  legend_position = "top",
  show_axis = c("x", "xy", "y", "none"),
  show_axis_text = c("xy", "x", "y", "none"),
  show_grid = c("y", "xy", "x", "none"),
  show_border = FALSE,
  x_text_angle = 0,
  x_text_hjust = NULL,
  ...
)
}
\description{
Basic ggplot2 theme
}
