% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataType-Triangle.R
\docType{data}
\name{Triangle}
\alias{Triangle}
\alias{calc.Triangle}
\alias{calc.Triangle.ChainLadder}
\alias{calc.Triangle.BF}
\alias{.Helper}
\alias{CapeCod}
\alias{tri_latest}
\alias{tri_link_ratio}
\alias{ata}
\title{Create a Triangle object}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
Triangle(x, origin = 1L)

## Develope a Triangle using Chain Ladder method
calc(Triangle, "ChainLadder", link_ratio = "weighted")

## Develope a Triangle using B-F method
calc(Triangle, "BF", expected_loss, link_ratio = "weighted")

## Helper Functions

CapeCod(x, premium, link_ratio = "weighted", ...)

tri_latest(tri)

tri_link_ratio(x, average = c(NA, "weighted", "simple"))

ata(tri, link_ratio, tail_factor = 1, floor = 0, interp = NULL)
}
\description{
Create a Triangle object
}
\examples{
prem <- Triangle(TRI$Premium)
loss <- Triangle(TRI$Incurred)
loss / prem

## Chain-Ladder Method
calc(prem, "ChainLadder", "weighted")
calc(prem, "ChainLadder", c(1.2, 1.1, 1.05))
calc(prem, "ChainLadder", "weighted", floor = 1)

## BF with CapeCod ELR
UltPrem <- calc(prem, "ChainLadder", floor = 1)$ultimate
ELR <- CapeCod(loss, premium = UltPrem, floor = 1)
calc(loss, "BF", expected_loss = ELR * UltPrem, floor = 1)
}
\keyword{datasets}
