% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func-ContractTerm.R
\name{XL}
\alias{XL}
\title{Helper function to create a XL contract structure}
\usage{
XL(
  x = NULL,
  Limit = Inf,
  Deductible = 0,
  AAL = Inf,
  AAD = 0,
  Reinstatement = NA_character_,
  Premium = 0
)
}
\description{
Helper function to create a XL contract structure
}
\examples{
XL(c("100M xs 200M", "200M xs 300M"))
XL(c("100M xs 200M xs 50M", "200M xs 300M"))
XL(Limit = 100, Deductible = 200, Reinstatement = "2@100\%")
XL(Limit = 100, Deductible = c(100, 200))
}
