% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opre-package.R
\docType{package}
\name{opre-package}
\alias{opre}
\alias{opre-package}
\title{opre: An R toolkit for reinsurance loss modeling and pricing}
\description{
`opre` is an R toolkit designed to accelarate modeling and pricing process in reinsurance transactions
}
\author{
\strong{Maintainer}: Charles Pan \email{addnox@gmail.com}

}
\keyword{internal}
